Refactor: Consolidate user update logic into `updateUser` service
This commit refactors the user update logic by consolidating the `edit`, `approveInfluencer`, and `declineInfluencer` functions into a single `updateUser` function within the User service. This simplifies the code and reduces duplication.  The controller now calls `updateUser` with the appropriate parameters to achieve the desired updates. Also updated the request types